// modules/loan/modules/b2b-payment/enum/b2b-payment-control-name.enum.ts
export enum B2bPaymentKeys {
  NAME = 'name',
  DUE_DATE = 'dueDate',
  BALANCE = 'balance',
  B2BID = 'b2bID',
}

export enum B2bPaymentForm {
  ACCOUNT_DEBITED = 'accountDebited',
  ACCOUNT = 'no_back',
  NAME = 'name',
  STATUS = 'status',
  CURRENCY = 'currency',
  INTEREST_RATE = 'interestRate',
  DEFAULT_RATE = 'defaultRate',
  OPENING_DATE = 'openingDate',
  EXPIRATIONS_DATE = 'expirationDate',
  INTEREST_BALANCE = 'interestBalance',
  BALANCE_AFTER_PAYMENT = 'balanceAfterPayment',
  CAPITAl = 'capital',
  INTEREST = 'interest',
  FCI = 'fci',
  ARREARS = 'arrears',
  BALANCE_PAYMENT = 'paymentBalance',
}

export enum PaymentTableType {
  PAY = 'payment',
}